invoice_mvp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pytest>=8.0.0
httpx>=0.27.0
openai>=1.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
